Repaired task06
